+++ RunningRacesApi/Services/ICsvExportService.cs
@@ -1,0 +1,6 @@
+﻿namespace RunningRacesApi.Services;
+
+public interface ICsvExportService
+{
+    string Export<T>(IEnumerable<T> data, IEnumerable<string> columns);
+}